Ensure webpack bundle analyzer works
