+++ webpack/envs/index.js
@@ -2,3 +2,4 @@ export * from "./base"
 export * from "./development"
 export * from "./ci"
 export * from "./production"
+export * from "./debug"
